refactor(aruba): co-locate typed network Ref constructors with their resource files
Move each XxxRef constructor (and its test) from the grouping files
pkg/aruba/refs.go and refs_test.go into the matching resource_xxx.go /
resource_xxx_test.go. Mirrors the one-file-per-resource convention used
across the wrapper layer. No public surface change; pure relocation.

Refs #268.
