+++ pkg/aruba/resource_security_group_test.go
@@ -914,3 +914,15 @@ func TestSecurityGroupsClientAdapter_Get_InjectsRefresh(t *testing.T) {
 		t.Error("Get should inject a refresh callback into the returned SecurityGroup")
 	}
 }
+
+func TestSecurityGroupRef(t *testing.T) {
+	ref := SecurityGroupRef("p-1", "vpc-1", "sg-1")
+	want := "/projects/p-1/providers/Aruba.Network/vpcs/vpc-1/securitygroups/sg-1"
+	if ref.URI() != want {
+		t.Errorf("SecurityGroupRef URI = %q, want %q", ref.URI(), want)
+	}
+	ids := parseURIIDs(ref.URI())
+	if ids["projects"] != "p-1" || ids["vpcs"] != "vpc-1" || ids["securitygroups"] != "sg-1" {
+		t.Errorf("parseURIIDs = %v", ids)
+	}
+}
