+++ pkg/aruba/resource_subnet_test.go
@@ -1034,3 +1034,15 @@ func TestSubnetsClientAdapter_Get_InjectsRefresh(t *testing.T) {
 		t.Error("Get should inject a refresh callback into the returned Subnet")
 	}
 }
+
+func TestSubnetRef(t *testing.T) {
+	ref := SubnetRef("p-1", "vpc-1", "sn-1")
+	want := "/projects/p-1/providers/Aruba.Network/vpcs/vpc-1/subnets/sn-1"
+	if ref.URI() != want {
+		t.Errorf("SubnetRef URI = %q, want %q", ref.URI(), want)
+	}
+	ids := parseURIIDs(ref.URI())
+	if ids["projects"] != "p-1" || ids["vpcs"] != "vpc-1" || ids["subnets"] != "sn-1" {
+		t.Errorf("parseURIIDs = %v", ids)
+	}
+}
