+++ pkg/aruba/resource_vpc_peering_test.go
@@ -1019,3 +1019,15 @@ func TestVPCPeeringsClientAdapter_Get_InjectsRefresh(t *testing.T) {
 		t.Error("Get should inject a refresh callback into the returned VPCPeering")
 	}
 }
+
+func TestVPCPeeringRef(t *testing.T) {
+	ref := VPCPeeringRef("p-1", "vpc-1", "peer-1")
+	want := "/projects/p-1/providers/Aruba.Network/vpcs/vpc-1/vpcPeerings/peer-1"
+	if ref.URI() != want {
+		t.Errorf("VPCPeeringRef URI = %q, want %q", ref.URI(), want)
+	}
+	ids := parseURIIDs(ref.URI())
+	if ids["projects"] != "p-1" || ids["vpcs"] != "vpc-1" || ids["vpcPeerings"] != "peer-1" {
+		t.Errorf("parseURIIDs = %v", ids)
+	}
+}
