+++ pkg/aruba/resource_vpn_tunnel_test.go
@@ -1348,3 +1348,15 @@ func TestVPNTunnelsClientAdapter_Get_InjectsRefresh(t *testing.T) {
 		t.Error("Get should inject a refresh callback into the returned VPNTunnel")
 	}
 }
+
+func TestVPNTunnelRef(t *testing.T) {
+	ref := VPNTunnelRef("p-1", "tun-1")
+	want := "/projects/p-1/providers/Aruba.Network/vpnTunnels/tun-1"
+	if ref.URI() != want {
+		t.Errorf("VPNTunnelRef URI = %q, want %q", ref.URI(), want)
+	}
+	ids := parseURIIDs(ref.URI())
+	if ids["projects"] != "p-1" || ids["vpnTunnels"] != "tun-1" {
+		t.Errorf("parseURIIDs = %v", ids)
+	}
+}
